# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SSB HRD API",
        "description": "API documentation for HRD modules in SSB Project",
        "contact": {
            "email": "admin@ssb.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://hrd.ptssb.my.id/api",
            "description": "SSB API Server"
        }
    ],
    "paths": {
        "/hrd/profile/{id}": {
            "get": {
                "tags": [
                    "HRD"
                ],
                "summary": "Get employee profile by ID",
                "operationId": "6fa1272793f40e61614fdc80eac6a55c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Employee ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/hrd/employee/department/{departmentId}": {
            "get": {
                "tags": [
                    "HRD"
                ],
                "summary": "Get employees by department ID (optional)",
                "operationId": "81e9c5742d0772b98c2af7df6bdc9383",
                "parameters": [
                    {
                        "name": "departmentId",
                        "in": "path",
                        "description": "Department ID. Set to 0 to get all employees.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/hrd/photo/{id}": {
            "get": {
                "tags": [
                    "HRD"
                ],
                "summary": "Get employee photo by ID",
                "operationId": "ec81c405151784298964edfd74d5841d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Employee ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "headers": {
                            "Content-Type": {
                                "description": "image/jpeg",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "Photo not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/hrd/memo/{id}": {
            "get": {
                "tags": [
                    "HRD"
                ],
                "summary": "Get internal memo document by ID",
                "operationId": "babe668715f608145884d19b3c75f772",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Memo ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "headers": {
                            "Content-Type": {
                                "description": "application/pdf",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access or inactive memo"
                    },
                    "404": {
                        "description": "Memo not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login dan dapatkan Bearer token",
                "description": "Autentikasi user menggunakan NIK dan password. Token yang dihasilkan digunakan sebagai Bearer token untuk endpoint yang memerlukan autentikasi.",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nik",
                                    "password"
                                ],
                                "properties": {
                                    "nik": {
                                        "description": "Nomor Induk Karyawan",
                                        "type": "string",
                                        "example": "001"
                                    },
                                    "password": {
                                        "description": "Password user",
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|AbCdEf..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nik": {
                                                    "type": "string",
                                                    "example": "001"
                                                },
                                                "nm": {
                                                    "type": "string",
                                                    "example": "Budi Santoso"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "NIK atau password salah"
                    },
                    "403": {
                        "description": "User tidak memiliki role"
                    }
                }
            }
        },
        "/refresh-token": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh access token",
                "description": "Revoke the current token and issue a new one",
                "operationId": "a2c5ee3fb1631cf5aa178730dfd8bee5",
                "responses": {
                    "200": {
                        "description": "New token issued",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|abc123..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}